+++ common/src/main/java/com/github/cao/awa/annuus/information/compressor/deflate/DeflateCompressor.java
@@ -11,7 +11,6 @@ import java.util.zip.InflaterInputStream;
 
 public class DeflateCompressor implements InformationCompressor {
 
-
     private final int compressLevel;
     private final String name;